Move deployment result types to router-types
