+++ src/lib/user-deployments/router-types.ts
@@ -5,14 +5,24 @@ import type { Deployment, DeploymentBuild } from '@/db/schema';
 import type { Event, BuildStatus } from './types';
 import type { EnvVarResponse } from './env-vars-validation';
 import type { GetPasswordStatusResponse } from './dispatcher-client';
-import type { RenameDeploymentResult, CheckSlugAvailabilityResult } from './deployments-service';
-
 /**
  * Result type for creating a deployment - discriminated union
  */
 export type CreateDeploymentResult =
   | { success: true; deploymentId: string; deploymentSlug: string; deploymentUrl: string }
   | { success: false; error: 'payment_required' | 'invalid_slug' | 'slug_taken'; message: string };
+
+export type RenameDeploymentResult =
+  | { success: true; deploymentUrl: string }
+  | {
+      success: false;
+      error: 'not_found' | 'invalid_slug' | 'slug_taken' | 'internal_error';
+      message: string;
+    };
+
+export type CheckSlugAvailabilityResult =
+  | { available: true }
+  | { available: false; reason: 'invalid_slug' | 'slug_taken'; message: string };
 
 /**
  * Represents an owner that can be either a user or an organization
